# distsql: refine code for setting LimitSize of kvRequest